Refactor test data and update test imports
Moved PDF test data files from tests/unit/pdf_data to tests/unit/helper and updated references in test files accordingly. PDF files are now stored under reports/pdf. Removed redundant and legacy test files, and simplified test imports to use the new helper structure. Updated CONTRIBUTING.md to reflect new test data and PDF file locations.
