+++ tests/unit/helper/hyperlinks-test.ts
