+++ tests/unit/helper/image-test.ts
