+++ tests/unit/helper/solar-energy.ts
