+++ tests/unit/test-02-get-text/label-test-pdf-text.test.ts
@@ -1,9 +1,9 @@
-import { data } from '../pdf_data/label-test';
 import { writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
+import { PDFParse } from 'pdf-parse';
 import { describe, expect, test } from 'vitest';
 
-import { PDFParse } from '../../src/index';
+import { data } from '../helper/label-test';
 
 const __pdf_txt = join(__dirname, data.textFile);
