+++ tests/unit/test-02-get-text/password-123456.test.ts
@@ -1,8 +1,9 @@
-import { data } from '../pdf_data/password-123456';
 import { writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import { PDFParse } from 'pdf-parse';
 import { describe, expect, test } from 'vitest';
+
+import { data } from '../helper/password-123456';
 
 const __pdf_txt = join(__dirname, data.textFile);
