+++ tests/unit/test-02-get-text/solar-energy-pdf-text.test.ts
@@ -1,9 +1,9 @@
-import { data } from '../pdf_data/solar-energy';
 import { writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
+import { PDFParse } from 'pdf-parse';
 import { describe, expect, test } from 'vitest';
 
-import { PDFParse } from '../../src/index';
+import { data } from '../helper/solar-energy';
 
 describe('test-02 all:true', async () => {
 	const buffer = await data.getBuffer();
