--- tests/unit/test-dist-builds/package-esm.test.ts
@@ -1,19 +1,0 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import { fileURLToPath } from 'node:url';
-import { PDFParse } from 'pdf-parse';
-import { describe, expect, test } from 'vitest';
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
-
-describe('test-dist local cjs test', async () => {
-	test('default export', async () => {
-		const file = path.join(__dirname, '../pdf_file/full-test.pdf');
-		const data = fs.readFileSync(file);
-
-		const parser = new PDFParse({ data });
-		const result = await parser.getText();
-
-		expect(result.text).toBeTruthy();
-	});
-});
