+++ tests/unit/test-example/url.test.ts
@@ -2,7 +2,7 @@ import { PDFParse } from 'pdf-parse';
 import { describe, expect, test } from 'vitest';
 
 describe('test-load url string', () => {
-	test('results be correct', { timeout: 10000 }, async () => {
+	test('results be correct', { timeout: 15000 }, async () => {
 		const parser = new PDFParse({ url: 'https://bitcoin.org/bitcoin.pdf' });
 		const infoResult = await parser.getInfo();
 		const textResult = await parser.getText();
@@ -14,7 +14,7 @@ describe('test-load url string', () => {
 });
 
 describe('test-load url object', () => {
-	test('results be correct', { timeout: 10000 }, async () => {
+	test('results be correct', { timeout: 15000 }, async () => {
 		const parser = new PDFParse({ url: new URL('https://bitcoin.org/bitcoin.pdf') });
 		const infoResult = await parser.getInfo();
 		const textResult = await parser.getText();
