+++ tests/unit/test-geometry/geometry.test.ts
@@ -1,8 +1,8 @@
 import { describe, expect, test } from 'vitest';
 
-import { Line, LineDirection } from '../../src/geometry/Line';
-import { LineStore } from '../../src/geometry/LineStore';
-import { Point } from '../../src/geometry/Point';
+import { Line, LineDirection } from '../../../src/geometry/Line';
+import { LineStore } from '../../../src/geometry/LineStore';
+import { Point } from '../../../src/geometry/Point';
 
 describe('basic geometry test', (): void => {
 	test('line constructor', (): void => {
